refactor: rework NamespaceConfig service.
The changes herein rework the NamespaceConfig service definition so
that adding a namespace config and updating a relation within a
namespace config are two separate methods. This helps us ensure that
relations are not overwritten (therefore deleted). We don't support
deletion of relations within a namespace config at this time, so this
helps eliminate that concern.