fix: disallow references to undefined relations and misc (#17)
The changes herein disallow references in relation tuples to undefined
relations in namespace configs. If a write is issued that references a
relation that is undefined at that particular snapshot in time, an error
is returned. Other changes include exponential retry-backoff for namespace
monitoring and more appropriate error status responses.